use native graphic library instead of cairo

+++ ui/row.go
@@ -871,8 +871,8 @@ func (me *row) Draw(canvas nux.Canvas) {
 			cms := cs.MeasuredSize()
 			if d, ok := child.(nux.Draw); ok {
 				canvas.Save()
-				canvas.Translate(cms.Position.Left, cms.Position.Top)
-				canvas.ClipRect(0, 0, cms.Width, cms.Height)
+				canvas.Translate(float32(cms.Position.Left), float32(cms.Position.Top))
+				canvas.ClipRect(0, 0, float32(cms.Width), float32(cms.Height))
 				d.Draw(canvas)
 				canvas.Restore()
 			}
